Initial commit for codeql sast
Initial commit for security static code analysis tooling via GitHub Advanced Security leveraging CodeQL as the engine.

Note that this is configured to run on `workflow_dispatch` and not `push` or `pull_request`. We are looking to validate the behavior of the scanner first.

See #proj-github-adv-sec-int - https://britecore.slack.com/archives/C01M6LMQ05Q for additional detail.